Document heatmap and reorganization
Added documentation for heatmap class.
Heatmap and histogram moved to the Components folder (this folder contains the panels that conforms SimViz UI)
